starlark import working as per jareds wizardry
