+++ .gitignore
@@ -1,20 +1,20 @@
 # Ignore all bazel-* symlinks. There is no full list since this can change
 # based on the name of the directory bazel is cloned into.
-/bazel-*
+/**/bazel-*
 
 # Ignore backup files.
-*~
+/**/*~
 
 # Ignore Vim swap files.
-.*.swp
+/**/.*.swp
 
 # Ignore files generated by IDEs.
-/.classpath
-/.factorypath
-/.idea/
-/.ijwb/
-/.project
-/.settings
+/**/.classpath
+/**/.factorypath
+/**/.idea/
+/**/.ijwb/
+/**/.project
+/**/.settings
 
 # Ignore mac related files.
-/.DS_Store
+/**/.DS_Store
