+++ DEVELOPER.md
@@ -1,0 +1,5 @@
+# For Developers
+
+bazel query 'kind(java_library,//src/main/java/net/starlark/...)'
+
+bazel query @maven//... | grep <package_youre_looking_for>
