--- client_vscode/BUILD
@@ -1,7 +1,0 @@
-load("@build_bazel_rules_nodejs//:index.bzl", "npm_install")
-
-npm_install(
-    name = "npm",
-    package_json = ":package.json",
-    package_lock_json = ":package-lock.json",
-)
