+++ client_vscode/package.json
@@ -379,17 +379,15 @@
         "vscode:prepublish": "./scripts/build.sh",
         "watch": "./scripts/build.sh -watch",
         "pretest": "npm run compile && npm run lint",
-        "test": "node ./out/src/test/run_test.js"
+        "test": "node ./out/src/test/run.js"
     },
     "devDependencies": {
-        "@bazel/bazelisk": "^1.7.3",
-        "@bazel/ibazel": "^0.14.0",
         "@types/glob": "^7.1.3",
         "@types/mocha": "^8.0.0",
         "@types/node": "^6.14.9",
         "@types/vscode": "^1.30.0",
         "glob": "^7.1.6",
-        "mocha": "^8.1.3",
+        "mocha": "^8.2.1",
         "tslint": "^5.20.1",
         "typescript": "^3.7.2",
         "vscode-test": "^1.3.0"
