+++ server/WORKSPACE
@@ -1,0 +1,71 @@
+workspace(
+    name = "bazel_language_server",
+    managed_directories = { 
+        "@npm": ["node_modules"],
+    },
+)
+
+load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
+load("@bazel_tools//tools/build_defs/repo:git.bzl", "git_repository")
+
+###############################
+## MAVEN - JAVA DEPENDENCIES ##
+###############################
+
+RULES_JVM_EXTERNAL_TAG = "3.3"
+RULES_JVM_EXTERNAL_SHA = "d85951a92c0908c80bd8551002d66cb23c3434409c814179c0ff026b53544dab"
+
+http_archive(
+    name = "rules_jvm_external",
+    strip_prefix = "rules_jvm_external-%s" % RULES_JVM_EXTERNAL_TAG,
+    sha256 = RULES_JVM_EXTERNAL_SHA,
+    url = "https://github.com/bazelbuild/rules_jvm_external/archive/%s.zip" % RULES_JVM_EXTERNAL_TAG,
+)
+
+load("@rules_jvm_external//:defs.bzl", "maven_install")
+
+maven_install(
+    artifacts = [
+        # guava
+        "com.google.guava:guava:29.0-jre",
+
+        # gson
+        "com.google.code.gson:gson:2.8.6",
+
+        # junit
+        "junit:junit:4.12",
+
+        # lsp4j
+        "org.eclipse.lsp4j:org.eclipse.lsp4j:0.10.0",
+        "org.eclipse.lsp4j:org.eclipse.lsp4j.debug:0.10.0",
+        "org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc:0.10.0",
+        "org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc.debug:0.10.0",
+        "org.eclipse.lsp4j:org.eclipse.lsp4j.generator:0.10.0",
+        "org.eclipse.lsp4j:org.eclipse.lsp4j.websocket:0.10.0",
+
+        # log4j
+        "org.apache.logging.log4j:log4j-core:2.13.3",
+        "org.apache.logging.log4j:log4j-api:2.13.3",
+
+        # mockito
+        "org.mockito:mockito-core:3.5.15",
+        "org.mockito:mockito-junit-jupiter:3.5.15",
+    ],
+    repositories = [
+        "https://repo1.maven.org/maven2",
+        "https://jcenter.bintray.com/",
+        "https://maven.google.com",
+    ],
+)
+
+#######################################
+## BAZELBUILD - JAVA STARLARK PARSER ##
+#######################################
+
+git_repository(
+    name = "bazel",
+    remote = "https://github.com/BYU-Bazel/bazel.git",
+    branch = "dependable-packages",
+)
+
+# TODO: Add in the bazel starlark parser
